requirements.txt
streamlit
google-gernativeai
python-dotenv
langchain
PyPDF2
chromadb